requirements.txt
tensorflow~=2.17.0
numpy~=1.26.0
opencv-python~=4.9.0.80
PyQt5~=5.15.10
pillow~=10.3.0
scipy~=1.13.0
tqdm~=4.66.4
PyWavelets~=1.8.0
torch>=2.6
torchvision